% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_preds.R
\name{scale_preds}
\alias{scale_preds}
\title{Scale Predictors}
\usage{
scale_preds(dt)
}
\arguments{
\item{dt}{a data.table (e.g., \link{train} or \link{test})}
}
\value{
Returns a list of length 3: the data.table with scaled predictors, a vector of the means in \code{dt}, a vector of the standard deviations in \code{dt}
}
\description{
Scales the predictors in the training or test data set
}
\details{
subtracts mean then divides by standard deviation
}

