% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test}
\alias{test}
\title{Test Data Set}
\format{data.table}
\usage{
test
}
\description{
Test data set from the Facebook 5 Kaggle competition
}
\details{
Need to predict unknown \code{place_id}'s for this data set.
}
\seealso{
\link{train}
}
\keyword{datasets}

